+++ pkg/networkserver/utils.go
@@ -19,12 +19,14 @@ import (
 	"fmt"
 
 	"go.thethings.network/lorawan-stack/v3/pkg/band"
+	"go.thethings.network/lorawan-stack/v3/pkg/component"
 	"go.thethings.network/lorawan-stack/v3/pkg/events"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	. "go.thethings.network/lorawan-stack/v3/pkg/networkserver/internal"
 	"go.thethings.network/lorawan-stack/v3/pkg/networkserver/internal/time"
 	"go.thethings.network/lorawan-stack/v3/pkg/networkserver/mac"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+	"go.thethings.network/lorawan-stack/v3/pkg/unique"
 )
 
 // nsScheduleWindow returns minimum time.Duration between downlink being added to the queue and it being sent to GS for transmission.
@@ -381,9 +383,63 @@ func (ns *NetworkServer) enqueueApplicationUplinks(ctx context.Context, ups ...*
 	if n == 0 {
 		return
 	}
-	logger := log.FromContext(ctx).WithField("uplink_count", n)
-	logger.Debug("Enqueue application uplinks for sending to Application Server")
-	if err := ns.applicationUplinks.Add(ctx, ups...); err != nil {
-		logger.WithError(err).Warn("Failed to enqueue application uplinks for sending to Application Server")
+
+	enqueue := func(ctx context.Context, ups ...*ttnpb.ApplicationUp) {
+		log.FromContext(ctx).Debug("Enqueue application uplinks for sending to Application Server")
+		if err := ns.applicationUplinks.Add(ctx, ups...); err != nil {
+			log.FromContext(ctx).WithError(err).Warn("Failed to enqueue application uplinks for sending to Application Server")
+		}
 	}
+
+	send := func(ctx context.Context, appID ttnpb.ApplicationIdentifiers, ups ...*ttnpb.ApplicationUp) error {
+		conn, err := ns.GetPeerConn(ctx, ttnpb.ClusterRole_APPLICATION_SERVER, &appID)
+		if err != nil {
+			log.FromContext(ctx).WithError(err).Warn("Failed to get Application Server peer")
+			return err
+		}
+		if err := ns.sendApplicationUplinks(ctx, ttnpb.NewNsAsClient(conn), appID, ups...); err != nil {
+			log.FromContext(ctx).WithError(err).Error("Failed to send application uplinks")
+			return err
+		}
+		return nil
+	}
+
+	partitionAndSend := func(ctx context.Context) error {
+		registerUplinkSenderStarted()
+		defer registerUplinkSenderFinished()
+
+		m := make(map[string][]*ttnpb.ApplicationUp)
+		for _, up := range ups {
+			appID := up.ApplicationIdentifiers.ApplicationId
+			m[appID] = append(m[appID], up)
+		}
+		for id, ups := range m {
+			appID := ttnpb.ApplicationIdentifiers{
+				ApplicationId: id,
+			}
+			ctx := log.NewContextWithFields(ctx, log.Fields(
+				"application_uid", unique.ID(ctx, appID),
+				"uplink_count", len(ups),
+			))
+			if err := send(ctx, appID, ups...); err != nil {
+				log.FromContext(ctx).WithError(err).Warn("Failed to send application uplinks to Application Server")
+				enqueue(ctx, ups...)
+			}
+		}
+		return nil
+	}
+
+	if !ns.uplinkQueueSemaphore.TryAcquire(1) {
+		enqueue(ctx, ups...)
+		return
+	}
+
+	ns.StartTask(&component.TaskConfig{
+		Context: ns.FromRequestContext(ctx),
+		ID:      sendApplicationUplinkTaskName,
+		Func:    partitionAndSend,
+		Done:    func() { ns.uplinkQueueSemaphore.Release(1) },
+		Restart: component.TaskRestartNever,
+		Backoff: component.DialTaskBackoffConfig,
+	})
 }