+++ CHANGELOG.md
@@ -17,6 +17,7 @@ For details about compatibility between different releases, see the **Commitment
 
 - Cache Root CA for client TLS configuration.
 - Identity Server no longer allows removing the `_ALL` right from entity collaborators if that leaves the entity without any collaborator that has the `_ALL` right.
+- The Network Server application uplink queue may now be skipped if the Application Server peer is available at enqueue time.
 
 ### Deprecated